Bugfix: date formatting
This fixes a bunch of issues to do with the formatting of dates. Previously we had a bunch of custom filters, some of which relied on other libraries such as luxon and lodash.

These have now all been removed, and the [date filters](https://x-govuk.github.io/govuk-prototype-filters/date/) from a plugin are used instead.

3 new filters were added to extract the separate date fields from an ISO date string, for use in the interview edit screen.